--- house-automation-frontend/src/app/core/guard/authenticated.guard.ts
@@ -1,28 +1,0 @@
-import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
-import { Observable } from 'rxjs';
-import { StorageKeys } from '../enums/storage-keys.enum';
-import { StorageService } from '../services/storage.service';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthenticatedGuard implements CanActivate {
-  constructor(
-    private storage: StorageService,
-    private router: Router,
-  ) { }
-
-  canActivate(
-    next: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    const token = this.storage.retrieve(StorageKeys.tokenKey);
-
-    if(!token) {
-      return this.router.createUrlTree(['/login']);
-    }
-    
-    return true;
-  }
-  
-}
