+++ house-automation-frontend/src/app/core/guard/authenticated/authenticated.guard.spec.ts
