+++ house-automation-frontend/src/app/layout/menu/menu.component.scss
@@ -1,6 +1,7 @@
 .app-menu {
   background-color: #3f51b5;
   padding: 16px;
+  box-shadow: 0 3px 5px -1px rgba(0,0,0,.2), 0 6px 10px 0 rgba(0,0,0,.14), 0 1px 18px 0 rgba(0,0,0,.12);
 
   & > .bar {
     display: flex;
