+++ house-automation-frontend/src/app/login/pages/login-page/login-form.scss
@@ -1,0 +1,15 @@
+.login-form {
+  box-shadow: 0 3px 1px -2px rgba(0,0,0,.2), 0 2px 2px 0 rgba(0,0,0,.14), 0 1px 5px 0 rgba(0,0,0,.12);
+  padding: 20px;
+  width: 40%;
+
+  & > .formfield {
+    display: block;
+    margin-bottom: 8px;
+  }
+
+  & > .button {
+    width: 100%;
+    margin-top: 20px;
+  }
+}
