+++ house-automation-frontend/src/styles.scss
@@ -4,4 +4,4 @@
 @import './assets/scss/generic/universal';
 @import './assets/scss/elements/typography';
 @import './assets/scss/elements/list';
-
+@import '@angular/material/prebuilt-themes/indigo-pink.css';
